Tomas api tests (#62)
* APITest working

* Fixed Tokons tests

* Merge with main

* Tested getComments

* merge main

* Wrote API tests for comments, likes, liked

* merge main

+++ backend/database/views.py
@@ -271,7 +271,7 @@ class CommentsAPIs(viewsets.ViewSet):
     def getComments(self, request, *args, **kwargs):
         postId = kwargs["postId"]
         if not Posts.objects.filter(id=postId).count() == 1:
-            return Response({"Tried to send non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"Tried to get comments from non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
 
         queryset = Comments.objects.filter(post_id=postId).order_by('-published')
         serializer = CommentsSerializer(queryset, many=True)
@@ -288,9 +288,9 @@ class CommentsAPIs(viewsets.ViewSet):
 
         #check that authorId and postId exist
         if not Authors.objects.filter(id=authorId).count() ==1:
-            return Response({"Tried to send post to non-existent author"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"Tried to make comment from non-existent author"}, status=status.HTTP_400_BAD_REQUEST)
         if not Posts.objects.filter(id=postId).count() == 1:
-            return Response({"Tried to send non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"Tried to make comment on non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
 
         author = Authors.objects.get(id=authorId)
         post = Posts.objects.get(id=postId)
@@ -299,13 +299,13 @@ class CommentsAPIs(viewsets.ViewSet):
         body = JSONParser().parse(io.BytesIO(request.body))
         contentType = 'text/plain'
         if 'contentType' in body:
-            if body['contentBody'] == 'text/markdown':
+            if body['contentType'] == 'text/markdown':
                 contentType = 'text/markdown'
-            elif body['contentBody'] != 'text/plain':
-                return Response({"Failed comment creation. Invalid input for contentBody."}, status=status.HTTP_400_BAD_REQUEST)
+            elif body['contentType'] != 'text/plain':
+                return Response({"Failed comment creation. Invalid input for contentType."}, status=status.HTTP_400_BAD_REQUEST)
 
-        if 'comment' in body:
-            comment = str(body['comment'])
+        if 'comment' in body and str(body['comment']).strip() != '':
+            comment = str(body['comment']).strip()
         else:
             return Response({"Failed comment creation. Missing 'comment' column."}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -395,7 +395,9 @@ class LikesAPIs(viewsets.ViewSet):
             return Response({"Tried to delete a like on a non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
         liker = Authors.objects.get(id=likerId)
         post = Posts.objects.get(id=postId)
-        
+
+        if Likes.objects.filter(post=post, author=liker).count() < 1:
+            return Response({"Tried to delete a non-existent like"}, status=status.HTTP_400_BAD_REQUEST)
         Likes.objects.filter(post=post, author=liker).delete()
         
         return Response({"Delete Like Successful"}, status=status.HTTP_200_OK)
@@ -415,8 +417,10 @@ class LikesAPIs(viewsets.ViewSet):
             return Response({"Tried to delete a like on a non-existent comment"}, status=status.HTTP_400_BAD_REQUEST)
         liker = Authors.objects.get(id=likerId)
         comment = Comments.objects.get(id=commentId)
-        
-        Likes.objects.filter(comment=comment, author=liker).delete()
+
+        if LikesComments.objects.filter(comment=comment, author=liker).count() < 1:
+            return Response({"Tried to delete a non-existent like"}, status=status.HTTP_400_BAD_REQUEST)
+        LikesComments.objects.filter(comment=comment, author=liker).delete()
         
         return Response({"Delete Like Successful"}, status=status.HTTP_200_OK)
 
@@ -433,6 +437,9 @@ class LikesAPIs(viewsets.ViewSet):
             size = int(request.GET.get('size',10))
         except:
             return Response("{Page or Size not an integer}", status=status.HTTP_400_BAD_REQUEST )
+
+        if Posts.objects.filter(id=postId).count() < 1:
+            return Response({"Tried to get likes from a non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
             
         likeObjs = []
         for like in Likes.objects.filter(post_id=postId):
@@ -459,6 +466,9 @@ class LikesAPIs(viewsets.ViewSet):
             size = int(request.GET.get('size',10))
         except:
             return Response("{Page or Size not an integer}", status=status.HTTP_400_BAD_REQUEST )
+
+        if Comments.objects.filter(id=commentId).count() < 1:
+            return Response({"Tried to get likes from a non-existent comment"}, status=status.HTTP_400_BAD_REQUEST)
 
         commentLikeObjs = []
         for like in LikesComments.objects.filter(comment_id=commentId):
@@ -509,6 +519,9 @@ class LikedAPIs(viewsets.ViewSet):
             size = int(request.GET.get('size',10))
         except:
             return Response("{Page or Size not an integer}", status=status.HTTP_400_BAD_REQUEST )
+
+        if not Authors.objects.filter(id=authorId).count() == 1:
+            return Response({"Tried to check likes of a non-existent author"}, status=status.HTTP_400_BAD_REQUEST)
 
         likeObjs = []
         for like in Likes.objects.filter(author_id=authorId):
